Add example test file for #1
Also changed project structure to remove superfluous `frege` folder.

Next thing is to put the project name in the module paths I
think... Plus I could do with more input on a suitable base program and
tests. I kinda like Fibonacci but it's hard to do property-based tests
on it :)